+++ lib/src/platform_app.dart
@@ -39,6 +39,8 @@ abstract class _BaseData {
     this.shortcuts,
     this.actions,
     this.onGenerateInitialRoutes,
+    this.highContrastDarkTheme,
+    this.highContrastTheme,
   });
 
   final Key widgetKey;
@@ -66,6 +68,8 @@ abstract class _BaseData {
   final Map<LogicalKeySet, Intent> shortcuts;
   final Map<Type, Action<Intent>> actions;
   final InitialRouteListFactory onGenerateInitialRoutes;
+  final ThemeData highContrastDarkTheme;
+  final ThemeData highContrastTheme;
 }
 
 class MaterialAppData extends _BaseData {
@@ -219,9 +223,6 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
   final Map<Type, Action<Intent>> actions;
   final InitialRouteListFactory onGenerateInitialRoutes;
 
-  final PlatformBuilder<MaterialAppData> android;
-  final PlatformBuilder<CupertinoAppData> ios;
-
   final PlatformBuilder2<MaterialAppData> material;
   final PlatformBuilder2<CupertinoAppData> cupertino;
 
@@ -252,18 +253,13 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
     this.shortcuts,
     this.actions,
     this.onGenerateInitialRoutes,
-    @Deprecated('Use material argument. material: (context, platform) {}')
-        this.android,
-    @Deprecated('Use cupertino argument. cupertino: (context, platform) {}')
-        this.ios,
     this.material,
     this.cupertino,
   }) : super(key: key);
 
   @override
   createMaterialWidget(BuildContext context) {
-    final data =
-        android?.call(context) ?? material?.call(context, platform(context));
+    final data = material?.call(context, platform(context));
 
     return MaterialApp(
       key: data?.widgetKey ?? widgetKey,
@@ -312,13 +308,14 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
       actions: data?.actions ?? actions,
       onGenerateInitialRoutes:
           data?.onGenerateInitialRoutes ?? onGenerateInitialRoutes,
+      highContrastDarkTheme: data?.highContrastDarkTheme,
+      highContrastTheme: data?.highContrastTheme,
     );
   }
 
   @override
   createCupertinoWidget(BuildContext context) {
-    final data =
-        ios?.call(context) ?? cupertino?.call(context, platform(context));
+    final data = cupertino?.call(context, platform(context));
 
     return CupertinoApp(
       key: data?.widgetKey ?? widgetKey,