+++ lib/src/platform_dialog_action.dart
@@ -51,6 +51,8 @@ class MaterialDialogActionData extends _BaseData {
     this.visualDensity,
     this.onLongPress,
     this.mouseCursor,
+    this.height,
+    this.minWidth,
   }) : super(
           widgetKey: widgetKey,
           child: child,
@@ -77,6 +79,8 @@ class MaterialDialogActionData extends _BaseData {
   final VisualDensity visualDensity;
   final VoidCallback onLongPress;
   final MouseCursor mouseCursor;
+  final double height;
+  final double minWidth;
 }
 
 class CupertinoDialogActionData extends _BaseData {
@@ -104,11 +108,6 @@ class PlatformDialogAction
   final Widget child;
   final VoidCallback onPressed;
 
-  @Deprecated('Use material argument. material: (context, platform) {}')
-  final PlatformBuilder<MaterialDialogActionData> android;
-  @Deprecated('Use cupertino argument. cupertino: (context, platform) {}')
-  final PlatformBuilder<CupertinoDialogActionData> ios;
-
   final PlatformBuilder2<MaterialDialogActionData> material;
   final PlatformBuilder2<CupertinoDialogActionData> cupertino;
 
@@ -117,15 +116,12 @@ class PlatformDialogAction
     this.widgetKey,
     @required this.child,
     @required this.onPressed,
-    this.android,
-    this.ios,
     this.material,
     this.cupertino,
   }) : super(key: key);
   @override
   FlatButton createMaterialWidget(BuildContext context) {
-    final data =
-        android?.call(context) ?? material?.call(context, platform(context));
+    final data = material?.call(context, platform(context));
 
     return FlatButton(
       key: data?.widgetKey ?? widgetKey,
@@ -151,13 +147,14 @@ class PlatformDialogAction
       visualDensity: data?.visualDensity,
       onLongPress: data?.onLongPress,
       mouseCursor: data?.mouseCursor,
+      height: data?.height,
+      minWidth: data?.minWidth,
     );
   }
 
   @override
   CupertinoDialogAction createCupertinoWidget(BuildContext context) {
-    final data =
-        ios?.call(context) ?? cupertino?.call(context, platform(context));
+    final data = cupertino?.call(context, platform(context));
 
     return CupertinoDialogAction(
       key: data?.widgetKey ?? widgetKey,