+++ lib/src/platform_text_field.dart
@@ -102,6 +102,9 @@ class MaterialTextFieldData {
     this.obscuringCharacter,
     this.autofillHints,
     this.mouseCursor,
+    this.onAppPrivateCommand,
+    this.cursorHeight,
+    this.restorationId,
   });
 
   final Key widgetKey;
@@ -151,6 +154,9 @@ class MaterialTextFieldData {
   final String obscuringCharacter;
   final Iterable<String> autofillHints;
   final MouseCursor mouseCursor;
+  final AppPrivateCommandCallback onAppPrivateCommand;
+  final double cursorHeight;
+  final String restorationId;
 }
 
 class CupertinoTextFieldData {
@@ -207,6 +213,8 @@ class CupertinoTextFieldData {
     this.selectionWidthStyle,
     this.obscuringCharacter,
     this.autofillHints,
+    this.cursorHeight,
+    this.restorationId,
   });
 
   final Key widgetKey;
@@ -261,14 +269,13 @@ class CupertinoTextFieldData {
   final ui.BoxWidthStyle selectionWidthStyle;
   final String obscuringCharacter;
   final Iterable<String> autofillHints;
+  final double cursorHeight;
+  final String restorationId;
 }
 
 class PlatformTextField
     extends PlatformWidgetBase<CupertinoTextField, TextField> {
   final Key widgetKey;
-
-  final PlatformBuilder<MaterialTextFieldData> android;
-  final PlatformBuilder<CupertinoTextFieldData> ios;
 
   final PlatformBuilder2<MaterialTextFieldData> material;
   final PlatformBuilder2<CupertinoTextFieldData> cupertino;
@@ -321,6 +328,9 @@ class PlatformTextField
   final String obscuringCharacter;
   final Iterable<String> autofillHints;
 
+  final double cursorHeight;
+  final String restorationId;
+
   PlatformTextField({
     Key key,
     this.widgetKey,
@@ -365,10 +375,8 @@ class PlatformTextField
     this.selectionWidthStyle,
     this.obscuringCharacter,
     this.autofillHints,
-    @Deprecated('Use material argument. material: (context, platform) {}')
-        this.android,
-    @Deprecated('Use cupertino argument. cupertino: (context, platform) {}')
-        this.ios,
+    this.cursorHeight,
+    this.restorationId,
     this.material,
     this.cupertino,
   })  : keyboardType = keyboardType ??
@@ -377,8 +385,7 @@ class PlatformTextField
 
   @override
   TextField createMaterialWidget(BuildContext context) {
-    final data =
-        android?.call(context) ?? material?.call(context, platform(context));
+    final data = material?.call(context, platform(context));
 
     return TextField(
       key: data?.widgetKey ?? widgetKey,
@@ -439,13 +446,15 @@ class PlatformTextField
       obscuringCharacter: data?.obscuringCharacter ?? obscuringCharacter ?? '•',
       autofillHints: data?.autofillHints ?? autofillHints,
       mouseCursor: data?.mouseCursor ?? MouseCursor.defer,
+      onAppPrivateCommand: data?.onAppPrivateCommand,
+      cursorHeight: data?.cursorHeight ?? cursorHeight,
+      restorationId: data?.restorationId ?? restorationId,
     );
   }
 
   @override
   CupertinoTextField createCupertinoWidget(BuildContext context) {
-    final data =
-        ios?.call(context) ?? cupertino?.call(context, platform(context));
+    final data = cupertino?.call(context, platform(context));
 
     return CupertinoTextField(
       key: data?.widgetKey ?? widgetKey,
@@ -517,6 +526,8 @@ class PlatformTextField
           ui.BoxWidthStyle.tight,
       obscuringCharacter: data?.obscuringCharacter ?? obscuringCharacter ?? '•',
       autofillHints: data?.autofillHints ?? autofillHints,
+      cursorHeight: data?.cursorHeight ?? cursorHeight,
+      restorationId: data?.restorationId ?? restorationId,
     );
   }
 }