+++ lib/src/platform_widget.dart
@@ -10,38 +10,30 @@ import 'platform.dart';
 import 'widget_base.dart';
 
 class PlatformWidget extends PlatformWidgetBase<Widget, Widget> {
-  final PlatformBuilder<Widget> android;
-  final PlatformBuilder<Widget> ios;
-
   final PlatformBuilder2<Widget> material;
   final PlatformBuilder2<Widget> cupertino;
 
   PlatformWidget({
     Key key,
-    @Deprecated('Use cupertino argument. cupertino: (context, platform) {}')
-        this.ios,
-    @Deprecated('Use material argument. material: (context, platform) {}')
-        this.android,
     this.cupertino,
     this.material,
   }) : super(key: key);
 
   @override
   Widget createMaterialWidget(BuildContext context) {
-    if (android == null && material == null) {
+    if (material == null) {
       return SizedBox.shrink();
     } else {
-      return android?.call(context) ??
-          material?.call(context, platform(context));
+      return material?.call(context, platform(context));
     }
   }
 
   @override
   Widget createCupertinoWidget(BuildContext context) {
-    if (ios == null && cupertino == null) {
+    if (cupertino == null) {
       return SizedBox.shrink();
     } else {
-      return ios?.call(context) ?? cupertino?.call(context, platform(context));
+      return cupertino?.call(context, platform(context));
     }
   }
 }